remove delete button
