+++ frontend/src/components/Employees/EmployeeListItem.jsx
@@ -1,18 +1,43 @@
 import React, { useContext } from "react";
-import { Td, Tr } from "@chakra-ui/react";
-import EditButton from "../Buttons/EditButton";
-import DeleteButton from "../Buttons/DeleteButton";
+import { IconButton, Td, Tr } from "@chakra-ui/react";
+import { DeleteIcon } from "@chakra-ui/icons";
 import { employeesContext } from "../../Providers/EmployeesProvider";
 import EditModal from "../Modal/EditModal";
 
 const EmployeeListItem = () => {
-    const { employees } = useContext(employeesContext);
+    const { employees, setEmployees } = useContext(employeesContext);
 
     const employeeList = employees?.map((employee) => {
         const salaryFormat = new Intl.NumberFormat("en-US", {
             style: "currency",
             currency: "USD",
         });
+
+        async function handleDeleteEvent(id) {
+            const employeeId = id;
+            console.log("ID: ", id);
+
+            if (window.confirm(`Are you sure you want to delete employee?`)) {
+                try {
+                    const url = `/api/employees/${employeeId}`;
+                    const response = await fetch(url, {
+                        method: "DELETE",
+                    });
+                    // const data = await response.json();
+                    // console.log("DELETED", data);
+
+                    const newEmployeeList = employees.filter((employee) => {
+                        return employee.id !== employeeId;
+                    });
+
+                    setEmployees(newEmployeeList);
+                } catch (err) {
+                    console.log(err);
+                }
+            } else {
+                console.log("Delete Aborted");
+            }
+        }
 
         return (
             <Tr key={employee.id}>
@@ -23,7 +48,12 @@ const EmployeeListItem = () => {
                     <EditModal id={employee.id} />
                 </Td>
                 <Td>
-                    <DeleteButton id={employee.id} />
+                    <IconButton
+                        colorScheme="red"
+                        size="sm"
+                        icon={<DeleteIcon />}
+                        onClick={() => handleDeleteEvent(employee.id)}
+                    />
                 </Td>
             </Tr>
         );
